Improvements made on filtering functionality for class schedule
Improvements made on filtering functionality for class schedule. User
can select class to see filtered from a dropdown menu on the calendar
page.

Next step is to remove orange background from empty table cells if the
class name has been removed and to replace the styling when the class
names are replaced should the user wish to perform a new filtering.