WOD area styling update. Implementation of automatically truncated workout description. Date display updated.
WOD area styling update. Now, there is separation between workout type
names (STRENGTH, WORKOUT, etc.) and description text.

Implementation of automatically truncated workout description. Text
added is capped to 25 words before the text is truncated and followed
by an ellipsis. The full workout is to appear on the workout page for
the given date.

Date display updated. The day is now in long form to add balance to the
section, e.g. “Thursday,” not “Thu” and “Saturday,” not “Sat.”