HTML and light styling applied for full blog and full workout page
HTML and light styling applied for full blog and full workout page
————————
The full blog page and the full workout page have the same format.

Under several archive examples will be a link to a full archive page
with pagination, cataloguing all workouts on the site.

**NOTE** Blog posts used to be posted below each workout. This is no
longer the case as the full blog and full workout pages are designed to
be separate pages, unlike the previous build.