Refactored default (non-media query) styling to rid app of "vh" elements
Refactored default (non-media query) styling to rid app of "vh"
elements so that the application is more responsive if the user’s
viewport is opened at different heights.

TODO - Refactor media query versions of the same elements so that the
responsive versions of the application behave the same as they did when
a “proper” viewport height was open as the width of the app changed.